feat(cc-widgets): fixed-styling
