+++ packages/contact-center/cc-components/src/components/task/CallControl/CallControlCustom/consult-transfer-popover-hooks.ts
@@ -1,7 +1,92 @@
 import {useCallback, useEffect, useRef, useState} from 'react';
 import {AddressBookEntry, ContactServiceQueue, EntryPointRecord, ILogger} from '@webex/cc-store';
 import {FetchPaginatedList} from '../../task.types';
-import {debounce, usePaginatedData} from './call-control-custom.utils';
+import {debounce} from './call-control-custom.utils';
+
+type TransformPaginatedData<T, U> = (item: T, page: number, index: number) => U;
+
+export const usePaginatedData = <T, U>(
+  fetchFunction: FetchPaginatedList<T> | undefined,
+  transformFunction: TransformPaginatedData<T, U>,
+  categoryName: string,
+  logger?: ILogger
+) => {
+  const [data, setData] = useState<U[]>([]);
+  const [page, setPage] = useState(0);
+  const [hasMore, setHasMore] = useState(true);
+  const [loading, setLoading] = useState(false);
+
+  const loadData = useCallback(
+    async (currentPage = 0, search = '', reset = false) => {
+      if (!fetchFunction) {
+        setData([]);
+        setHasMore(false);
+        return;
+      }
+
+      setLoading(true);
+      try {
+        const apiParams: {page: number; pageSize: number; search?: string} = {
+          page: currentPage,
+          pageSize: 25,
+        };
+
+        if (search && search.trim()) {
+          apiParams.search = search;
+        }
+
+        logger?.info(`CC-Components: Loading ${categoryName} - page: ${currentPage}, search: "${search}"`);
+        const response = await fetchFunction(apiParams);
+
+        if (!response || !response.data) {
+          logger?.error(`CC-Components: Invalid response from fetch function for ${categoryName}`);
+          if (reset || currentPage === 0) {
+            setData([]);
+          }
+          setHasMore(false);
+          return;
+        }
+
+        logger?.info(`CC-Components: Loaded ${response.data.length} ${categoryName} for page ${currentPage}`);
+
+        const transformedEntries = response.data.map((entry, index) => transformFunction(entry, currentPage, index));
+
+        if (reset || currentPage === 0) {
+          setData(transformedEntries);
+        } else {
+          setData((prev) => [...prev, ...transformedEntries]);
+        }
+
+        const newPage = response.meta?.page ?? currentPage;
+        const totalPages = response.meta?.totalPages ?? 1;
+
+        setPage(newPage);
+        setHasMore(newPage < totalPages - 1);
+
+        logger?.info(
+          `CC-Components: ${categoryName} pagination state - current: ${newPage}, total: ${totalPages}, hasMore: ${newPage < totalPages - 1}`
+        );
+      } catch (error) {
+        logger?.error(`CC-Components: Error loading ${categoryName}:`, error);
+        if (reset || currentPage === 0) {
+          setData([]);
+        }
+        setHasMore(false);
+      } finally {
+        setLoading(false);
+      }
+    },
+    [fetchFunction, transformFunction, logger, categoryName]
+  );
+
+  const reset = useCallback(() => {
+    setData([]);
+    setPage(0);
+    setHasMore(true);
+  }, []);
+
+  return {data, page, hasMore, loading, loadData, reset};
+};
 
 export type CategoryType = 'Agents' | 'Queues' | 'Dial Number' | 'Entry Point';
